Complete basic features

+++ app/src/main/java/com/example/rajitha/weatherapp/Model/Wind.java
@@ -4,7 +4,7 @@ public class Wind {
 
 
     private double speed ;
-    private int deg;
+    private float deg;
 
     public Wind() {
     }
@@ -18,7 +18,7 @@ public class Wind {
         this.speed = speed;
     }
 
-    public int getDeg() {
+    public float getDeg() {
         return deg;
     }
